% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2dist.R
\name{df2dist}
\alias{df2dist}
\title{A data frame to distance matrix function}
\usage{
df2dist(dat)
}
\arguments{
\item{dat}{Data frame to be converted to distance matrix.}
}
\description{
This function allows you to convert from a data frame
to a matrix of distance.
}
\examples{
df2dist()
}
\keyword{dataframe}
\keyword{distance}

